Enable OpenBOR in Emulationstation
OpenBOR only works with Pak files, if you have an extracted game you will need to create a pak first.

There is a bug/annoyance with the master.cfg file, if its the first time running a game it will start in a small window.

 You will need to go into options and change the video to full, after you do this, that game will always run in fullscreen.

 You need to do this the first time you run any pak.

 If master.cfg does not work, sound is weird or controller not working, you will need to use a keyboard to set your gamepad

 After you set up your gamne, copy the /storage/.config/openbor/Saves/{gamename}.cfg file to /storage/.config/openbor/master.cfg

master.cfg will only be copied the first time you run that particular game.

+++ packages/sx05re/sx05re/bin/sx05reRunEmu.sh
@@ -2,8 +2,10 @@
 
 hdmimode=`cat /sys/class/display/mode`;
 
-if  pgrep mpg123 >/dev/null ; then
-/storage/.emulationstation/scripts/bgm.sh 
+if [ "$3" != "KEEPMUSIC" ] || [ "$4" != "KEEPMUSIC" ]; then
+	if  pgrep mpg123 >/dev/null ; then
+	/storage/.emulationstation/scripts/bgm.sh 
+	fi
 fi
 
 # Set framebuffer geometry to match the resolution, splash should change according to the resolution. 
@@ -33,15 +35,21 @@ RUNTHIS='/usr/bin/retroarch -L /tmp/cores/${EMU}_libretro.so "$2"'
 
 # Else, read the first argument to see if its LIBRETRO, REICAST, MAME or PSP
 case $1 in
+"OPENBOR")
+	RUNTHIS='/usr/bin/openbor.sh "$2"'
+		;;
+"RETROPIE")
+	RUNTHIS='bash /retropie/scripts/fbterm.sh "$2"'
+		;;
 "LIBRETRO")
-   RUNTHIS='/usr/bin/retroarch -L /tmp/cores/$2_libretro.so "$3"'
-         ;;
+	RUNTHIS='/usr/bin/retroarch -L /tmp/cores/$2_libretro.so "$3"'
+		;;
 "REICAST")
-   RUNTHIS='/usr/bin/reicast.sh "$2"'
-         ;;
+	RUNTHIS='/usr/bin/reicast.sh "$2"'
+		;;
 "ADVMAME")
-   RUNTHIS='/usr/bin/advmame.sh "$2"'
-         ;;
+	RUNTHIS='/usr/bin/advmame.sh "$2"'
+		;;
 "PSP")
       if [ "$EMU" = "PPSSPPSA" ]; then
    #PPSSPP can run at 32BPP but only with buffered rendering, some games need non-buffered and the only way they work is if I set it to 16BPP
@@ -60,7 +68,9 @@ echo "Run Command is:" >> $SX05RELOG
 eval echo  ${RUNTHIS} >> $SX05RELOG 
 
 # TEMP: I need to figure out how to mix sounds, but for now make sure BGM is killed completely to free up the soundcard
-killall mpg123
+if [ "$3" != "KEEPMUSIC" ] || [ "$4" != "KEEPMUSIC" ]; then
+	killall mpg123 
+fi
 
 # Exceute the command and try to output the results to the log file if it was not dissabled.
 if [ "$3" == "NOLOG" ] || [ "$4" == "NOLOG" ]; then
@@ -71,13 +81,14 @@ else
    eval ${RUNTHIS} >> $SX05RELOG 2>&1
 fi 
 
+if [ "$3" != "KEEPMUSIC" ] || [ "$4" != "KEEPMUSIC" ]; then
+	DEFE=$(sed -n 's|\s*<bool name="BGM" value="\(.*\)" />|\1|p' /storage/.emulationstation/es_settings.cfg)
 
-DEFE=$(sed -n 's|\s*<bool name="BGM" value="\(.*\)" />|\1|p' /storage/.emulationstation/es_settings.cfg)
-
-if [ "$DEFE" == "true" ]; then
-killall mpg123
-/storage/.emulationstation/scripts/bgm.sh
-fi 
+	if [ "$DEFE" == "true" ]; then
+	killall mpg123
+	/storage/.emulationstation/scripts/bgm.sh
+	fi 
+fi
